Create subpackages for networking implementations
