+++ pkg/networking/cilium/testdata/cilium_manifest.yaml
