+++ pkg/networking/kindnetd/testdata/expected_kindnetd_manifest.yaml
