+++ pkg/networking/kindnetd/testdata/kindnetd_manifest.yaml
